* libm/math/w_atan2.c: Fix atan2 to be consistent with glibc         (atan2(+0,-0), atan2(-0,-0) and atan(-0,+0))         * libm/math/wf_atan2.c: Ditto.
        * libm/math/e_pow.c: Fix to be consistent with glibc with regards to
        treatment of NaN and +-inf arguments.
        * libm/math/ef_pow.c: Ditto.
        * libm/math/w_pow.c: Ditto.
        * libm/math/wf_pow.c: Ditto.
        * libm/math/w_acos.c: Fix domain errors to return NaN.
        * libm/math/w_asin.c: Ditto.
        * libm/math/wf_acos.c: Ditto.
        * libm/math/wf_asin.c: Ditto.
        * libm/math/w_log.c: Fix to return NaN for negative number inputs.
        * libm/math/wf_log.c: Ditto.
        * libm/math/wf_log10.c: Ditto.
        * libm/math/w_log10.c: Ditto.

+++ newlib/libm/math/w_pow.c
@@ -24,14 +24,14 @@ INDEX
 ANSI_SYNOPSIS
 	#include <math.h>
 	double pow(double <[x]>, double <[y]>);
-	float powf(float <[x]>, float <[y]>);
+	float pow(float <[x]>, float <[y]>);
 
 TRAD_SYNOPSIS
 	#include <math.h>
 	double pow(<[x]>, <[y]>);
 	double <[x]>, <[y]>;
 
-	float powf(<[x]>, <[y]>);
+	float pow(<[x]>, <[y]>);
 	float <[x]>, <[y]>;
 
 DESCRIPTION
